# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "3PS OpenAPI Documentation",
        "description": "3PS OpenAPI publish",
        "contact": {
            "email": "artexsdns@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.3permata.co.id/",
            "description": "3PS OpenAPI Published"
        }
    ],
    "paths": {
...
        "/api/v1/Customers": {
            "get": {
                "tags": [
                    "Customer View SO"
                ],
                "summary": "Get list of Sales Order Customer",
                "description": "Returns list of SO Customers",
                "operationId": "getTransportOrders",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransportOrders"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Transport_orders": {
                "title": "Transport_orders",
                "description": "Transport_orders model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 111
                    },
                    "ItemID_accurate": {
                        "title": "ItemID_accurate",
                        "description": "ItemID_accurate",
                        "type": "string",
                        "example": "CS0607203200271"
                    },
                    "company": {
                        "title": "Company",
                        "description": "Company",
                        "type": "string",
                        "format": "int64",
                        "example": "29"
                    },
                    "name": {
                        "title": "name",
                        "description": "Name",
                        "type": "string",
                        "format": "string",
                        "example": "PT. VIE LIE XANG"
                    },
                    "status_id": {
                        "title": "Status ID",
                        "description": "Author's id of the new project",
                        "type": "integer",
                        "format": "int64",
                        "example": 4
                    }
                },
                "type": "object"
            },
...
            "TransportOrders": {
                "title": "TransportOrders",
                "description": "Project resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Transport_orders"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ProjectResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}